mejoras en el codigo

+++ src/main/java/tuti/desi/servicios/CiudadServiceImpl.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Service;
 
 import tuti.desi.accesoDatos.ICiudadRepo;
 import tuti.desi.entidades.Ciudad;
+import tuti.desi.excepciones.EntidadNoEncontradaException;
 import tuti.desi.excepciones.Excepcion;
 import tuti.desi.presentacion.ciudades.CiudadesBuscarForm;
 
@@ -26,13 +27,10 @@ public class CiudadServiceImpl implements CiudadService {
 		return repo.findAll();
 	}
 
-
-
 	@Override
-	@Deprecated
 	public Ciudad getById(Long idCiudad) {
-
-		return repo.findById(idCiudad).get();
+		return repo.findById(idCiudad)
+				.orElseThrow(() -> new EntidadNoEncontradaException("la ciudad", idCiudad));
 	}
 	
 	@Override
@@ -40,35 +38,33 @@ public class CiudadServiceImpl implements CiudadService {
 	{
 		//ver https://docs.spring.io/spring-data/jpa/docs/1.5.0.RELEASE/reference/html/jpa.repositories.html
 		if(filter.getNombre()==null && filter.getProvinciaSeleccionada()==null)
-			//return repo.findAll();
-			throw new Excepcion("Es necesario al menos un filtro");
+			return repo.findAll();
+			//throw new Excepcion("Es necesario al menos un filtro");
 		else
 			return repo.findByNombreOrIdProvincia(filter.getNombre(),filter.getProvinciaSeleccionada());
 				
 	}
 
-
-
 	@Override
 	public void deleteByid(Long id) {
+		if (!repo.existsById(id)) {
+			throw new EntidadNoEncontradaException("la ciudad", id);
+		}
 		repo.deleteById(id);
-		
 	}
-
-
 
 	@Override
 	public void save(Ciudad c) throws Excepcion {
-		if(c.getId()==null && !repo.findByNombreAndIdProvincia(c.getNombre(), c.getProvincia().getId()).isEmpty()) //estoy dando de alta una nueva ciudad y ya existe una igual?
+		if(c.getId()==null 
+				&& !repo.findByNombreAndIdProvincia(c.getNombre(), c.getProvincia().getId()).isEmpty()) //estoy dando de alta una nueva ciudad y ya existe una igual?
 			throw new Excepcion("Ya existe una ciudad con el mismo nombre, para la misma provincia");  
 		else
 		{
-			if(!repo.findByNombreAndIdProvinciaAndIdNot(c.getNombre(), c.getProvincia().getId(),c.getId()).isEmpty()) //si edito el nombre, valido que no exista otra con el mismo nombre?
+			if(!repo.findOtraCiudadByNombreAndProvincia(c.getNombre(), c.getProvincia().getId(),c.getId()).isEmpty()) //si edito el nombre, valido que no exista otra con el mismo nombre?
 				throw new Excepcion("Existe otra ciudad con el mismo nombre para la misma provincia");
 			else
 				repo.save(c);
 		}
-		
 	}
 
 }
